move pdf service options
