+++ dfa-public/src/API/EMBC.DFA.PUBLIC.API/Services/ADFSTokenProvider.cs
@@ -79,14 +79,6 @@ namespace EMBC.DFA.API.Services
         public string ResourceName { get; set; }
     }
 
-    public class PdfServiceConfigs
-    {
-        public string GeneratePDFFile { get; set; }
-        public string ClientId { get; set; }
-        public string ClientSecret { get; set; }
-        public string TokenUrl { get; set; }
-    }
-
     public static class AccessTokenProviderEx
     {
         public static IServiceCollection AddADFSTokenProvider(this IServiceCollection services)
